Prevent virtual address allocated to uart from being reallocated later on
UART registers are memory mapped at physical address 0xe000.0000 in the SoC.
Before this commit I was mapping this paddr to vaddr 0xc800.0000.
But then later on this address was used by the vm subsystem and mapped
to another physical address which then produced a dead lock
when uart tried to check if the "send buffer" was empty.

In order to prevent re-allocation of the virtual address allocated for the uart
I am increasing the VM_MIN_KERNEL_ADDRESS (by 3 pages, 1 should be enough...)
So that 0xc800.0000 is never allocated by the vm sub system.