Add support for IRQ handling
We now save/restore IE upon exceptions.
IE should be shadowed in PSW but there is a bug in Qemu.
In Qemu any write to PSW will not update IE and any write
to IE will not write to PSW.
Therefore, until Qemu is fixed we need to save IE upon exceptions.