_do_real_tlb_miss_handling and __isr were called with IRQ enabled
If the TLB miss happens when IRQ are enabled, then
PSW.EIE is set, doing eret with this PSW will then enable back IRQ
(PSW.IE = PSW.EIE). We don't want to call the tls miss handler nor the irq
handler with irq enabled.

Therefore instead of reading PSW to modify it, we then start with a zero'ed register
and only add the bits we want to enable the TLB when calling _do_real_tlb_miss_handling
or __isr, and we end up writting this register to PSW.